# web/backend/routes/root.ts
import { FastifyInstance } from 'fastify'
import { JsonSchemaToTsProvider } from '@fastify/type-provider-json-schema-to-ts'
import { RuntimeApiClient } from '@platformatic/control'
import { modeSchema, pidParamSchema, selectableRuntimeSchema } from '../schemas'
import { getSelectableRuntimes } from '../utils/runtimes'
import { fileExists } from '../utils/files'
import { MS_WAITING } from '../utils/constants'

const wait = (ms: number) => new Promise(resolve => setTimeout(resolve, ms))

export default async function (fastify: FastifyInstance) {
  const typedFastify = fastify.withTypeProvider<JsonSchemaToTsProvider>()
  const api = new RuntimeApiClient()

  typedFastify.get('/runtimes', {
    schema: {
      querystring: {
        type: 'object',
        properties: {
          includeAdmin: {
            type: 'boolean',
            default: false,
          },
        },
      },
      response: { 200: { type: 'array', items: selectableRuntimeSchema } }
    }
  }, async (request) => {
    if (fastify.loaded.mode === 'load' && await fileExists(fastify.loaded.path)) {
      await wait(MS_WAITING)
      return fastify.loaded.runtimes
    }
    return getSelectableRuntimes(await api.getRuntimes(), request.query.includeAdmin)
  })

  typedFastify.get('/runtimes/:pid/health', {
    schema: {
      params: pidParamSchema,
      response: {
        200: {
          type: 'object',
          additionalProperties: false,
          properties: {
            status: {
              type: 'string',
              enum: ['OK', 'KO'],
              description: "Status can only be 'OK' or 'KO'"
            }
          },
          required: ['status']
        }
      }
    }
  }, async ({ params: { pid } }) => {
    const ok = { status: 'OK' as const }
    const ko = { status: 'KO' as const }

    if (fastify.loaded.mode === 'load') return ok
    try {
      const result = await api.getMatchingRuntime({ pid: pid.toString() })
      return (result.pid === pid) ? ok : ko
    } catch {
      return ko
    }
  })

  typedFastify.get('/runtimes/:pid/services', {
    schema: {
      params: pidParamSchema,
      response: {
        200: {
          type: 'object',
          additionalProperties: false,
          required: ['entrypoint', 'production', 'services'],
          properties: {
            entrypoint: {
              type: 'string'
            },
            production: {
              type: 'boolean'
            },
            services: {
              type: 'array',
              items: {
                anyOf: [
                  {
                    type: 'object',
                    required: ['id', 'type', 'status', 'version', 'localUrl', 'entrypoint', 'dependencies'],
                    properties: {
                      id: {
                        type: 'string'
                      },
                      type: {
                        type: 'string'
                      },
                      status: {
                        type: 'string'
                      },
                      version: {
                        type: 'string'
                      },
                      localUrl: {
                        type: 'string'
                      },
                      entrypoint: {
                        type: 'boolean'
                      },
                      workers: {
                        type: 'number'
                      },
                      url: {
                        type: 'string'
                      },
                      dependencies: {
                        type: 'array',
                        items: {
                          type: 'object',
                          required: ['id', 'url', 'local'],
                          properties: {
                            id: {
                              type: 'string'
                            },
                            url: {
                              type: 'string'
                            },
                            local: {
                              type: 'boolean'
                            }
                          }
                        }
                      }
                    }
                  },
                  {
                    type: 'object',
                    required: ['id', 'status'],
                    properties: {
                      id: {
                        type: 'string'
                      },
                      status: {
                        type: 'string'
                      }
                    }
                  }
                ]
              }
            }
          }
        }
      }
    }
  }, async (request) => {
    if (fastify.loaded.mode === 'load') {
      return fastify.loaded.services
    }
    return api.getRuntimeServices(request.params.pid)
  })

  typedFastify.get('/runtimes/:pid/openapi/:serviceId', {
    schema: {
      params: { type: 'object', properties: { pid: { type: 'number' }, serviceId: { type: 'string' } }, required: ['pid', 'serviceId'] }
    }
  }, async ({ params: { pid, serviceId } }) => {
    return api.getRuntimeOpenapi(pid, serviceId)
  })

  typedFastify.post('/runtimes/:pid/restart', {
    schema: { params: pidParamSchema, body: { type: 'object' } }
  }, async (request) => {
    try {
      await api.restartRuntime(request.params.pid)
    } catch (err) {
      fastify.log.warn({ err }, 'Issue restarting the runtime')
    }
  })

  typedFastify.get('/mode', {
    schema: {
      querystring: {
        type: 'object',
        properties: {
          all: {
            type: 'boolean',
            default: false,
          },
        },
      },
      response: {
        200: {
          type: 'object',
          additionalProperties: false,
          properties: {
            path: { type: 'string' },
            mode: modeSchema,
            runtimes: { type: 'array', items: selectableRuntimeSchema }
          },
          required: ['path', 'mode']
        }
      }
    }
  }, async ({ query: { all } }) => {
    const { loaded: { path, mode, runtimes } } = fastify
    return { path, mode, runtimes: all ? runtimes : undefined }
  })

  typedFastify.post('/mode', {
    schema: {
      body: {
        type: 'object',
        additionalProperties: false,
        properties: { mode: modeSchema },
        required: ['mode']
      }
    }
  }, async ({ body: { mode } }) => {
    fastify.loaded.mode = mode
  })
}